examples/portfoliomanager/pyproject.toml
[tool.hatch.build.targets.wheel]
packages = ["mcp_portfolio_server"]
[project]
name = "mcp-portfolio-manager"
version = "0.1.0"
description = "Portfolio management agent using MCP and OpenAI Agent SDK"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "MCP Python SDK" }]
keywords = ["mcp", "portfolio", "agent", "openai", "finance"]
license = { text = "MIT" }
dependencies = [
    "anyio>=4.5",
    "mcp",
    "openai>=1.0.0",
    "httpx>=0.27",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.6.9",
]

[project.scripts]
portfolio-server = "mcp_portfolio_server.server:main"
news-server = "mcp_news_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

